requirements.txt
google-cloud-vision==0.24.0
numpy==1.12.1
opencv-python==3.2.0
requests==2.14.2
scipy==0.19.0
